+++ CHANGELOG.md
@@ -1,3 +1,1 @@
-### Added
-- Added mob effect tooltip to /effect command completion text
-- Added a preview tooltip for paintings
+**Port to 26.3**